+++ src/factories/license.ts
@@ -17,11 +17,11 @@ SPDX-License-Identifier: Apache-2.0
 Copyright (c) OWASP Foundation. All Rights Reserved.
 */
 
-export * from './index.common'
+import {LicenseFactory as _LicenseFactory} from '../contrib/license/factories'
 
-// region node-specifics
-
-/** @deprecated next */
-export * as FromNodePackageJson from './fromNodePackageJson.node'
-
-// endregion node-specifics
+/**
+ * Deprecated — Alias of {@link Contrib.License.Factories.LicenseFactory}.
+ *
+ * @deprecated Use `Contrib.License.Factories.LicenseFactory` instead.
+ */
+export class LicenseFactory extends _LicenseFactory {}